requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
google-cloud-firestore
google-auth
google-auth-oauthlib
google-auth-httplib2
pandas
numpy
scikit-learn
xgboost
python-dotenv
